stop exporting constant time equals
