+++ lib/src/ml_dsa_base.dart
@@ -2,6 +2,7 @@
 
 import 'dart:typed_data';
 
+import 'comparison.dart';
 import 'conversion.dart';
 import 'entropy.dart';
 import 'expansion.dart';
@@ -316,11 +317,19 @@ class MLDSA {
         h = null;
       } else {
         final List<Int32List> ct0 = vectorNttInverse(
-            parameters, scalarVectorNtt(parameters, cHat, t0Hat));
-        final List<Int32List> ct0Neg =
-            scalarVectorMultiply(parameters, -1, ct0);
+          parameters,
+          scalarVectorNtt(parameters, cHat, t0Hat),
+        );
+        final List<Int32List> ct0Neg = scalarVectorMultiply(
+          parameters,
+          -1,
+          ct0,
+        );
         final List<Int32List> wPrime = vectorAddPolynomials(
-            parameters, vectorSubtractPolynomials(parameters, w, cs2), ct0);
+          parameters,
+          vectorSubtractPolynomials(parameters, w, cs2),
+          ct0,
+        );
 
         h = vectorMakeHint(parameters, ct0Neg, wPrime);
         final int ct0Max = vectorMaxAbsCoefficient(parameters, ct0);
@@ -384,23 +393,4 @@ class MLDSA {
     return zMax < (parameters.gamma1() - parameters.beta()) &&
         constantTimeEquals(cTilde, cTildePrime);
   }
-}
-
-bool constantTimeEquals(Uint8List a, Uint8List b) {
-  if (a.length != b.length) return false;
-
-  final len = a.length;
-  final aWords = Uint32List.view(a.buffer, a.offsetInBytes, len ~/ 4);
-  final bWords = Uint32List.view(b.buffer, b.offsetInBytes, len ~/ 4);
-
-  int result = 0;
-  for (int i = 0; i < aWords.length; i++) {
-    result |= aWords[i] ^ bWords[i];
-  }
-
-  for (int i = len & ~3; i < len; i++) {
-    result |= a[i] ^ b[i];
-  }
-
-  return result == 0;
 }
